Completed a test of Google/NCF sign in
Completed a test of signing in to an application with oauth. You can do this by going to the website and manually typing in the url ../oauth_test and login through there to get a confirmation screen. There currently is no logout so make sure to terminate the session, definently don't deploy to the server until further testing. Also note that a configuration file with the secret keys is needed, this will be given to the relevant people in private.