Updated dbhelper
Updated the DBHelper refactor to reflect the updates. Fixed error with constant names and reenabled tests, Jordan and I were mistaken about the casing upon further review. The DBTest was readded, we need this code to test  for the full version of the database helper since it isn't completely done yet. It is also my current reference for AOC structuring. It will be removed in the final version of DBHelper once the last methods are written